requirements.txt
streamlit==1.32.0
plotly==5.20.0
pandas==2.2.1
sqlalchemy==2.0.28
anthropic==0.21.3
python-dotenv==1.0.1
scikit-learn==1.4.1
openpyxl==3.1.2